Adding Category entity and it's service, config, dto, mapper and controller with crud apis #Category

+++ Server/ATA.Library.Server.Service/Contracts/IEntityService.cs
@@ -10,7 +10,7 @@ namespace ATA.Library.Server.Service.Contracts
 {
     public interface IEntityService<TEntity> where TEntity : class, IATAEntity, new()
     {
-        Task<TEntity?> GetByIdAsync(long key, CancellationToken cancellationToken);
+        Task<TEntity?> GetByIdAsync(int key, CancellationToken cancellationToken);
         IQueryable<TEntity> GetAll();
         IEnumerable<TEntity> Get(Expression<Func<TEntity, bool>>? where = null, Func<IQueryable<TEntity>, IOrderedQueryable<TEntity>>? orderBy = null);
         Task<IEnumerable<TEntity>> GetAsync(CancellationToken cancellationToken, Expression<Func<TEntity, bool>>? where = null, Func<IQueryable<TEntity>, IOrderedQueryable<TEntity>>? orderBy = null, string includeProperties = "");
@@ -21,7 +21,7 @@ namespace ATA.Library.Server.Service.Contracts
             CancellationToken cancellationToken);
 
         Task<TEntity> UpdateAsync(TEntity entity, CancellationToken cancellationToken);
-        Task DeleteAsync(long key, CancellationToken cancellationToken);
+        Task DeleteAsync(int key, CancellationToken cancellationToken);
         Task DeleteAsync(TEntity entity, CancellationToken cancellationToken);
     }
 }
